# Add Electron start-up performance script