--- a/.gitignore
+++ b/.gitignore
@@ -30,3 +30,4 @@ dev-packages/electron/compile_commands.json
 .eslintcache
 scripts/download
 license-check-summary.txt*
+*-trace.json