--- a/scripts/performance/.gitignore
+++ b/scripts/performance/.gitignore
@@ -3,3 +3,4 @@ workspace
 *.csv
 *.json
 !base-package.json
+!electron-trace-config.json