--- a/scripts/performance/base-package.json
+++ b/scripts/performance/base-package.json
@@ -1,31 +1,37 @@
 {
   "private": true,
-  "name": "@theia/example-browser",
-  "version": "1.18.0",
+  "name": "@theia/example-{{app}}",
+  "version": "{{version}}",
   "license": "EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0",
   "theia": {
+    "target": "{{app}}",
     "frontend": {
       "config": {
-        "applicationName": "Theia Browser Example",
+        "applicationName": "Theia {{app}} Example",
         "preferences": {
           "files.enableTrash": false
         }
       }
+    },
+    "backend": {
+      "config": {
+        "resolveSystemPlugins": false
+      }
     }
   },
   "dependencies": {
-    "@theia/core": "1.18.0",
-    "@theia/plugin-ext": "1.18.0"
+    "@theia/core": "{{version}}",
+    "@theia/plugin-ext": "{{version}}"
   },
   "scripts": {
     "clean": "theia clean",
     "build": "yarn compile && yarn bundle",
     "bundle": "theia build --mode development",
     "compile": "tsc -b",
-    "rebuild": "theia rebuild:browser --cacheRoot ../..",
+    "rebuild": "theia rebuild:{{app}} --cacheRoot ../..",
     "start": "yarn rebuild && THEIA_CONFIG_DIR=./theia-config-dir theia start --plugins=local-dir:../../noPlugins --log-level=fatal"
   },
   "devDependencies": {
-    "@theia/cli": "1.18.0"
+    "@theia/cli": "{{version}}"
   }
 }