… scripts/performance/README.md
… scripts/performance/extension-impact.js