@cdamus 2021-11-25
Thanks, @colin-grant-work . Good points, all. I've pushed commit 60f5cfd02031403ca24b717947f3d996103573d8 that should address them, including adding a `--debug` option for logging error output from the Theia child process (I did find I often needed this to diagnose issues). This works together with detection of the common case of the native modules being built for the wrong target to exit the script early.